requirements.txt
uvicorn==0.13.3
aiofiles==0.5.0
fastapi==0.61.1
pydantic==1.6.1
pymongo==3.11.0
starlette==0.13.6
jinja2==2.11.2
requests==2.25.1
pytest-mock==3.5.1
pytest-cov==2.11.1